pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "specode"
version = "0.1.0"
description = "A terminal-first AI chat assistant built with Typer, Rich, and Pydantic AI."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic-ai",
    "prompt-toolkit",
    "python-dotenv",
    "rich",
    "typer",
]

[project.scripts]
specode = "specode.cli:main"

[dependency-groups]
dev = [
    "pytest",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["src/specode"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
